Initial FUSE support!
It took a really long time to track down that waiting for spawn_mount
never worked, but also that once I moved to a separate thread I (obviously,
in retrospect) had to wait a bit for the filesystem to be ready.

Now the basic tests at least have getattr calls and readdir and so on. I've
sketched out the basic structs and mappings, but next up is to use fill them in
on init() then use them.