Initial create, write, flush, and release.
This is a first attempt at doing writes. The code isn't
prepared for anything other than a 'fresh create', but
it should be able to stream a large file.

GCS "requires" (currently) that ResumableUploads be written in
multiples-of-256 KiB chunks, except for the finalizing chunk. To do
so, I made the file handles in the FUSE layer have a
ResumableUploadCursor from the GCS interop layer. It went okay.