Check generation and meta-generation on reads.
GCS has two "serial numbers" for objects:

 - Generation
 - Metageneration

If someone overwrites an Object and doesn't have object versioning
enabled, the Generation changes. If someone "overwrites" an Object,
but has versioning enabled, the Generation stays the same but the
Metageneration changes.

We want to make sure that if someone is reading data, that they're
reading the thing they intended to. So, we can just check that both
Generation and Metageneration match the "cached" data in our Fileystem
mappings.

This commit just bubbles up the error code, which will cause reads to
*fail* if the underlying has been edited out of band. It's probably
better in the Filesystem code to explicitly notice that there is a
Generation/Metageneration mismatch and "do the right thing" (e.g., for
some filehandle like systems, it probably makes sense to allow the
open file handle to continue reading from the *previous* data,
assuming object versioning is enabled).