Remove reqwest entirely.
reqwest was helpful in getting started, but is sort of an awkward
liability compared to controlling our HTTP requests more directly
with hyper. The auth token fetching was the last user.

Now the only call into auth.rs is for adding the auth headers
(optionally), and I've updated the signatures to be async.