Fixup the warnings in the Filesystem code.
I did a lot of "let mut" on things that didn't need it. A real "bug"
was found: I wasn't mounting read-only as "-o ro" anymore. Fixing that
involved a slightly annoying game of building the list of options in
two parts.

Given this and the previous commit, we're now rustc-warning free. Next
up: do a pass fixing up clippy warnings.