Add direct large reads to the GCS tests.
I only had large reads being issued on the Filesystem side,
which makes it annoying to track down whether or not we're
*able* to issue reads and writes of the correct sizes.

wchargin@ also pointed out that I had a bug if a caller asked
for how_many == 0 (fixed, thanks!). And I left a long note
about overfetching versus not (I kind of prefer to let GCS
sort it out; it's happy to trim the read).