Add a mount_gcsfuser binary.
To make this easier to play with, I wanted to have a `mount`
compatible mount_gcsfuse (which on Linux should be mount.gcsfuse)
to take a GCS path and a local mountpoint before calling mount().

Tested it out with both a public bucket (the landsat one, including
a prefix) and a private bucket. As an example:

cargo run -- gs://gcp-public-data-landsat/LC08/01/044/034/ /tmp/my-mount -o ro

to mount the subdirectory of gcp-public-data-landsat to /tmp/my-mount as
read-only.